func-prog/memoization/pentagonal-numbers: complete solution, using Long

+++ functional-programming/memoization-and-dp/pentagonal-numbers/pentagonal-numbers.scala
@@ -1,9 +1,9 @@
 object Solution {
     import scala.collection.mutable.Map
 
-    val cache: Map[Int, Int] = Map(1 -> 1, 2 -> 5)
+    val cache: Map[Long, Long] = Map(1L -> 1L, 2L -> 5L)
 
-    def pentagonal(num: Int): Int = cache.getOrElseUpdate(num, 2 * num + (num - 2) + pentagonal(num - 1))
+    def pentagonal(num: Long): Long = cache.getOrElseUpdate(num, 2 * num + (num - 2) + pentagonal(num - 1))
 
     def main(args: Array[String]) {
         val cases = readInt
